+++ src/app/contact/contact.component.spec.ts
@@ -16,13 +16,31 @@
  */
 
 import { ContactComponent } from './contact.component';
-import { MockBuilder, MockedComponentFixture, MockRender } from 'ng-mocks';
+import { MockBuilder, MockedComponentFixture, MockRender, MockService } from 'ng-mocks';
+import { NGXLogger } from 'ngx-logger';
+import { AuthService } from '../services';
+import { Subject } from 'rxjs';
+import { MaxHeightContentContainerComponent } from '@gematik/demis-portal-core-library';
+import { FooterComponent } from '../footer/footer.component';
 
 describe('ContactComponent', () => {
   let fixture: MockedComponentFixture<ContactComponent>;
   let component: ContactComponent;
+  const tokenChangedSubject = new Subject<void>();
 
-  beforeEach(() => MockBuilder(ContactComponent));
+  beforeEach(() =>
+    MockBuilder(ContactComponent)
+      .keep(MaxHeightContentContainerComponent)
+      .keep(FooterComponent)
+      .mock(NGXLogger)
+      .provide({
+        provide: AuthService,
+        useValue: {
+          ...MockService(AuthService),
+          $tokenChanged: tokenChangedSubject,
+        } as AuthService,
+      })
+  );
 
   beforeEach(() => {
     fixture = MockRender(ContactComponent);